Add files via upload
This project uses BeautifulSoup4, request, and re modules in order to search CNBC's live currency exchange website. The code takes the currency name in a 3 letter code (USD) and returns a formatted string that shows the exchange rate, the rate change (both absolute and % change), and the date last updated.